print the folder id if installing the handler fails
the logging lacked the information which was the problematic folder id
even though the information is send to the syncthing process.
in combination with an upstream fix in notify ( https://github.com/Zillode/notify/pull/2 )
this will make debugging failures much easier as folder id and the failing node (i.e. path)
will be part of the error message.